Remove clone usage in demo by copying state handles (#169)
* Remove clone usage in demo by copying state handles

* Remove redundant clones and locals after Copy handles

* Fix clippy warnings across tests and demos

+++ crates/compose-core/src/tests/slot_backend_tests.rs
@@ -862,7 +862,7 @@ fn test_chunked_anchor_rebuild_after_shift() {
     let _g = storage.begin_group(1400);
     let mut ids = Vec::new();
     for i in 0..8 {
-        let s = storage.alloc_value_slot(|| i as i32);
+        let s = storage.alloc_value_slot(|| i);
         ids.push(s);
     }
     storage.end_group();
